brainwise_django/requirements.txt
asgiref==3.8.1
Django==4.2
dj-rest-auth==2.2.5
django-allauth==0.50.0
djangorestframework-simplejwt==5.0.0
django-cors-headers==4.2.0
django-fsm==3.0.0
djangorestframework==3.14.0
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.10.0
sqlparse==0.5.3
psycopg2-binary==2.9.7
